Window and graphics
Added GLEW and GLFW for window generation and graphics.
Made a GPU class to work with all the graphic logic.
Connected the GPU class to memory so the opcodes can access it.
Added Application class to hold our whole runtime logic, made it run a frame every time it loops for easy pause functionality.